requirements.txt
arrow==1.2.3
boto3==1.26.87
celery==5.3.6
cryptography==39.0.2
Flask==3.0.2
flower==2.0.1
gunicorn==22.0.0
pillow==10.3.0
pydantic==1.10.6
pydash==7.0.7
PyJWT==2.6.0
pymongo==4.3.3
python-dotenv==1.0.1
python-magic==0.4.27
redis==5.0.3
requests==2.31.0
selenium==4.18.1
Werkzeug==3.0.1
